file-based: new Permissions Reader Class to handle specific tasks of getting ACLs and Identities
